Break util func into a new file
